# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Docusign APIs in Node Express backend service",
      "version": "1.0.0",
      "title": "Docusign API",
      "contact": {
        "email": "info@crediation.com"
      }
    },
    "securityDefinitions": {
      "Bearer": {
        "description": "Docusign APIs allows a valid JWT token as an authentication method which must be passed in all the queries in the 'Authorization' header.\n The following syntax must be used in the 'Authorization' header:\n Bearer xxxxxx.yyyyyyy.zzzzzz",
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "host": "localhost:3001",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "produces": ["application/json"],
    "consumes": ["application/json"],
    "paths": {
      "/templates": {
        "get": {
          "summary": "Get all templates",
          "security": [
            {
              "Bearer": []
            }
          ],
          "responses": {
            "200": {
              "description": "Get all account's templates",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Template"
                }
              }
            }
          }
        }
      },
      "/envelopes": {
        "post": {
          "summary": "Create an envelope",
          "security": [
            {
              "Bearer": []
            }
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Envelope object to be created.",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Envelope"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Envelope created",
              "schema": {
                "type": "object",
                "properties":{
                  "data": {
                    "type": "object",
                    "properties": {
                      "envelopeId": {"type": "string", "example": "f3620881-3d18-406e-9c2a-0c1de321808a"},
                      "uri": {"type": "string", "example": "/envelopes/f3620881-3d18-406e-9c2a-0c1de321808a"},
                      "statusDateTime": {"type": "string", "example": "2020-08-17T09:30:43.5000000Z"},
                      "status": {"type": "string", "example": "sent"}
                    }
                  } 
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not found"
            },
            "415": {
              "description": "Unsupported Media Type"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "definitions": {
      "Envelope": {
        "type": "object",
        "properties": {
          "emailSubject": {"type": "string", "example": "Please sign the loan contract"},
          "status": {"type": "string", "example": "sent"},
          "templateId": {"type": "string", "example": "3e14c269-a599-4c2d-9350-b6ae7c0b2e80"},
          "templateRoles": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Roles"
            }
          }
        }
      },
      "Template": {
        "type": "object",
        "properties": {
          "templateId": {"type": "string"},
          "name": {"type": "string"},
          "shared": {"type": "string"},
          "password": {"type": "string"},
          "description": {"type": "string"},
          "lastModified": {"type": "string"},
          "created": {"type": "string"},
          "uri": {"type": "string"},
          "folderName": {"type": "string"},
          "folderId": {"type": "string"},
          "folderUri": {"type": "string"},
          "owner": {"$ref": "#/definitions/Owner"},
          "emailSubject": {"type": "string"},
          "emailBlurb": {"type": "string"},
          "signingLocation": {"type": "string"},
          "authoritativeCopy": {"type": "string"},
          "enforceSignerVisibility": {"type": "string"},
          "enableWetSign": {"type": "string"},
          "allowMarkup": {"type": "string"},
          "allowReassign": {"type": "string"}
        } 
      },
      "Owner": {
        "type": "object",
        "properties": {
          "userName": {"type": "string"},
          "userId": {"type": "string"},
          "email": {"type": "string"}
        }
      },
      "Roles": {
        "type": "object",
        "properties": {
          "email": {"type": "string", "example": "user@crediation.com"},
          "name": {"type": "string", "example": "Sample User"},
          "roleName": {"type": "string", "example": "applicant"}
        }
      }
    }
  }